python/domain-ssl-checker-cloudfunction/requirements.txt
functions-framework==3.*
requests==2.31.0
PyYAML>=5.4
